BugFix: Initial Booking Creation fix. Room, Accommodation, BookChannel, Person CRUD fixed. 	modified:   Controllers/AccommodationController.cs 	modified:   Controllers/PersonController.cs 	modified:   Controllers/RoomController.cs 	modified:   Hubs/CreateDocumentsHub.cs 	modified:   Views/Person/Edit.cshtml 	modified:   Views/Room/Delete.cshtml 	modified:   Views/Room/Details.cshtml 	modified:   Views/Room/Edit.cshtml 	modified:   bin/Debug/net9.0/MAppBnB.dll 	modified:   bin/Debug/net9.0/MAppBnB.exe 	modified:   bin/Debug/net9.0/MAppBnB.pdb 	modified:   bin/Debug/net9.0/MAppBnB.staticwebassets.endpoints.json 	modified:   bin/Debug/net9.0/MAppBnB.staticwebassets.runtime.json 	modified:   obj/Debug/net9.0/MAppBnB.AssemblyInfo.cs 	modified:   obj/Debug/net9.0/MAppBnB.AssemblyInfoInputs.cache 	modified:   obj/Debug/net9.0/MAppBnB.csproj.FileListAbsolute.txt 	modified:   obj/Debug/net9.0/MAppBnB.dll 	modified:   obj/Debug/net9.0/MAppBnB.pdb 	modified:   obj/Debug/net9.0/MAppBnB.sourcelink.json 	modified:   obj/Debug/net9.0/apphost.exe 	deleted:    obj/Debug/net9.0/compressed/2kv4t9g0ij-d135fqncnz.gz 	new file:   obj/Debug/net9.0/compressed/2kv4t9g0ij-lyw3kl5rsv.gz 	new file:   obj/Debug/net9.0/compressed/hlvuidnl8f-q6njfu2ecv.gz 	deleted:    obj/Debug/net9.0/compressed/hlvuidnl8f-vq708yxl3e.gz 	modified:   obj/Debug/net9.0/ref/MAppBnB.dll 	modified:   obj/Debug/net9.0/refint/MAppBnB.dll 	modified:   obj/Debug/net9.0/staticwebassets.build.endpoints.json 	modified:   obj/Debug/net9.0/staticwebassets.build.json 	modified:   obj/Debug/net9.0/staticwebassets.development.json 	modified:   obj/Debug/net9.0/staticwebassets/msbuild.MAppBnB.Microsoft.AspNetCore.StaticWebAssetEndpoints.props 	modified:   obj/Debug/net9.0/staticwebassets/msbuild.MAppBnB.Microsoft.AspNetCore.StaticWebAssets.props 	modified:   wwwroot/css/site.css 	modified:   wwwroot/js/searchPerson.js 	modified:   obj/Debug/net9.0/MAppBnB.GeneratedMSBuildEditorConfig.editorconfig 	modified:   obj/Debug/net9.0/MAppBnB.assets.cache 	modified:   obj/MAppBnB.csproj.nuget.dgspec.json 	modified:   obj/project.nuget.cache

+++ Controllers/PersonController.cs
@@ -10,6 +10,8 @@ using MAppBnB.Data;
 
 namespace MAppBnB.Controllers
 {
+    //TODO: se host, non farlo vedere nella lista delle persone. Qui e nell'hub.
+    //TODO: se è capofamiglia, capogruppo, o ospite singolo, documento obbligatorio
     public class PersonController : Controller
     {
         private readonly MappBnBContext _context;
@@ -22,7 +24,6 @@ namespace MAppBnB.Controllers
         // GET: Person
         public async Task<IActionResult> Index()
         {
-            ViewData["DocumentList"] = await _context.Document.ToListAsync();
             return View(await _context.Person.ToListAsync());
         }
 
@@ -209,8 +210,11 @@ namespace MAppBnB.Controllers
             var person = await _context.Person.FindAsync(id);
             if (person != null)
             {
-                var document = await _context.Document.FirstAsync(x => x.id == person.DocumentID);
-                _context.Document.Remove(document);
+                if (person.DocumentID != null)
+                {
+                    var document = await _context.Document.FirstOrDefaultAsync(x => x.id == person.DocumentID);
+                    _context.Document.Remove(document);
+                }
                 _context.Person.Remove(person);
             }
